Split parser enum hygiene guards
